IPL data analysis_2008_2017
The analysis is done on Dataset with all detailed information on IPL matches held from Year 2008 to Year 2017. Data contains Date, Venue, Toss winner, Match winner, Winning margin, Player of the match and Umpire for the match.

Objective of analysis is extract and understand the statistical overview of all these matches.